Support additional_type for the full range up to 27
substring() doesn't support bigint values,
but nonetheless we still want to try,
since the data_value even though being passed
in 8 bytes, could when decoded become a value
which fits in an integer. There is no real harm
in trying, since the user will just get a different
error message, "ERROR:  integer out of range"
instead of "ERROR: Decoding of CBOR type not implemented".

The rationale is that there might be implementations
that pass lengths as 4 or 8 bytes, even though the length values
when decoded are of lesser values than the bytes the occupy.