Add randomized tests covering the full range of possible values for major types
So far, tests for major types 0, 1 and 2 have been implemented in this way.

All possible values will not be tested for of course,
only as many as controlled using the \set iterations param,
but all values *have a chance* of being tested,
which means every time we run the tests, some new previously
untested values hopefully gets tested.